+++ simple_sim/src/world.rs
@@ -99,7 +99,7 @@ impl World {
             y,
             self.grid.height()
         );
-        self.grid.get(x as i64, y as i64)
+        self.grid.get(x, y)
     }
 
     pub fn width(&self) -> f32 {
